Aadhar_Masking/requirements.txt
fastapi
uvicorn
pillow
PyMuPDF
ollama
numpy
pandas
